+++ packages/x-adapter-platform/src/mappers/requests/ai/index.ts
@@ -1,3 +1,4 @@
 export * from './questions-request.mapper'
 export * from './suggestions-request.mapper'
+export * from './suggestions-search-request.mapper'
 export * from './tasks-request.mapper'