+++ packages/x-adapter-platform/src/schemas/responses/ai/suggestions-search-response.schema.ts
@@ -1,0 +1,13 @@
+import type { AiSuggestionsSearchResponse } from '@empathyco/x-types'
+import type { PlatformAiSuggestionsSearchResponse } from '../../../types/responses'
+import { createMutableSchema } from '@empathyco/x-adapter'
+import { aiSuggestionsSearchSchema } from '../../models/ai/suggestions-search.schema'
+
+/**
+ * Default implementation for the AIOverviewSuggestionsSearchResponseSchema.
+ * @public
+ */
+export const aiSuggestionsSearchResponseSchema = createMutableSchema<
+  PlatformAiSuggestionsSearchResponse,
+  AiSuggestionsSearchResponse
+>(aiSuggestionsSearchSchema)