+++ packages/x-adapter-platform/src/types/models/ai/index.ts
@@ -1,2 +1,1 @@
-export * from './question.model'
 export * from './suggestion-search.model'