+++ packages/x-adapter-platform/src/types/requests/ai/index.ts
@@ -1,3 +1,4 @@
 export * from './questions-request.model'
 export * from './suggestions-request.model'
+export * from './suggestions-search-request.model'
 export * from './tasks-request.model'