+++ packages/x-adapter-platform/src/types/responses/ai/index.ts
@@ -1,2 +1,3 @@
 export * from './questions-response.model'
+export * from './suggestions-search-response.model'
 export * from './tasks-response.model'